Move operators to their own file
The list of operators could grow and it should be a more modular
architecture to have them defined in separate files.
